refactor: Keep open-search as default client
